Added components folder and moved App.js component to components
