+++ src/components/App.js
@@ -1,0 +1,12 @@
+import logo from './logo.svg';
+import './App.css';
+
+function App() {
+  return (
+    <div>
+      
+    </div>
+  );
+}
+
+export default App;
